Style nbarlow part of site
