--- Pages/Users/Nbarlow/Home.razor
@@ -1,11 +1,0 @@
-@page "/users/nbarlow"
-
-@if (_self == null)
-{
-    <p><em>Loading...</em></p>
-}
-else
-{
-    <h1>Hey it's me, @(_self.FullName)!</h1>
-    <h2>Check out my <a href="/users/nbarlow/how-old-am-i">birthday calculator!</a></h2>
-}
