--- Pages/Users/Nbarlow/Home.razor.cs
@@ -1,29 +1,0 @@
-
-using ExoKomodo.Models;
-using Microsoft.AspNetCore.Components;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Net.Http;
-using System.Net.Http.Json;
-using System.Threading.Tasks;
-
-namespace ExoKomodo.Pages.Users.Nbarlow
-{
-    public partial class Home
-    {
-        [Inject]
-        private HttpClient _http { get; set; }
-        private const string UserId = "nbarlow";
-        private User _self;
-
-        protected override async Task OnInitializedAsync()
-        {
-            _self = (await _http.GetFromJsonAsync<List<User>>("data/users.json")).Where(user => user.Id == UserId).FirstOrDefault();
-            if (_self is null)
-            {
-                throw new Exception($"Could not find user {UserId}");
-            }
-        }
-    }
-}
