+++ Pages/Users/Nbarlow/Index.razor.cs
@@ -1,0 +1,89 @@
+using Microsoft.JSInterop;
+using ExoKomodo.Config;
+using ExoKomodo.Models;
+using Microsoft.AspNetCore.Components;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Net.Http;
+using System.Net.Http.Json;
+using System.Threading.Tasks;
+using ExoKomodo.Pages.Users.Jorson;
+
+namespace ExoKomodo.Pages.Users.Nbarlow
+{
+    internal class IndexBase : PageBase {}
+
+    public partial class Index : IDisposable
+    {
+        #region Public
+
+        #region Constructors
+        public Index()
+        {
+            _base = new IndexBase();   
+        }
+        #endregion
+
+        #region Constants
+        public const string UserId = "nbarlow";
+        #endregion
+
+        #endregion
+
+        #region Protected
+
+        #region Member Methods
+        protected override void OnAfterRender(bool firstRender)
+        {
+            base.OnAfterRender(firstRender);
+
+            if (firstRender)
+            {
+                _base.Initialize();
+            }
+        }
+
+        protected override async Task OnInitializedAsync()
+        {
+            _self = (await _http.GetFromJsonAsync<List<User>>("data/users.json")).Where(user => user.Id == UserId).FirstOrDefault();
+            if (_self is null)
+            {
+                throw new Exception($"Could not find user {UserId}");
+            }
+        }
+        #endregion
+
+        #endregion
+
+        #region Private
+
+        #region Members
+        [Inject]
+        private HttpClient _http { get; set; }
+        private bool _isDisposed { get; set; }
+        private PageBase _base { get; set; }
+        private User _self;
+        #endregion
+
+        #endregion
+
+        #region IDisposable Support
+        public void Dispose()
+        {
+            Dispose(true);
+            GC.SuppressFinalize(this);
+        }
+
+        protected virtual void Dispose(bool disposing)
+        {
+            if (_isDisposed || !disposing)
+            {
+                return;
+            }
+            _base.Dispose();
+            _isDisposed = true;
+        }
+        #endregion
+    }
+}
