+++ wwwroot/css/nbarlow/home.css
@@ -1,0 +1,11 @@
+.creeper-crunch-me {
+    background-image: url("/img/nbarlow/creeper-crunch-me.jpg");
+    background-repeat: round;
+    background-size: contain;
+    height: 30rem;
+    width: 100%;
+}
+
+.content {
+    font-size: 1.5rem;
+}
