+++ wwwroot/css/nbarlow/how-old-am-i/how-old-am-i.css
@@ -1,0 +1,4 @@
+input#birthdate {
+    border-radius: 5px;
+    border-color: #d69503;
+}
